Split off Orient-specific code in OrientTestDb
